requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-autoslug==1.9.8
django-s3-sqlite==0.0.5
django-tinymce==3.5.0
idna==3.4
instamojo-wrapper==1.2.0
MouseInfo==0.1.3
mysqlclient==2.1.1
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.3.0
tzdata==2022.4
urllib3==1.26.12